Fix GeometryReader uses
