+++ MedoDelirioBrasilia/Sources/Views/Reactions/Detail/ReactionDetailView.swift
@@ -48,72 +48,75 @@ struct ReactionDetailView: View {
     // MARK: - View Body
 
     var body: some View {
-        ScrollView {
-            VStack {
-                ContentList(
-                    viewModel: soundListViewModel,
-                    soundSearchTextIsEmpty: .constant(nil),
-                    showNewTag: false,
-                    dataLoadingDidFail: viewModel.dataLoadingDidFail,
-                    reactionId: viewModel.reaction.id,
-                    headerView: {
-                        ReactionDetailHeader(
-                            title: viewModel.reaction.title,
-                            subtitle: viewModel.subtitle,
-                            imageUrl: viewModel.reaction.image,
-                            attributionText: viewModel.reaction.attributionText,
-                            attributionURL: viewModel.reaction.attributionURL
+        GeometryReader { geometry in
+            ScrollView {
+                VStack {
+                    ContentList(
+                        viewModel: soundListViewModel,
+                        soundSearchTextIsEmpty: .constant(nil),
+                        showNewTag: false,
+                        dataLoadingDidFail: viewModel.dataLoadingDidFail,
+                        reactionId: viewModel.reaction.id,
+                        containerSize: geometry.size,
+                        headerView: {
+                            ReactionDetailHeader(
+                                title: viewModel.reaction.title,
+                                subtitle: viewModel.subtitle,
+                                imageUrl: viewModel.reaction.image,
+                                attributionText: viewModel.reaction.attributionText,
+                                attributionURL: viewModel.reaction.attributionURL
+                            )
+                            .frame(height: 260)
+                            .padding(.bottom, 6)
+                        },
+                        loadingView: LoadingView(),
+                        emptyStateView: EmptyStateView(
+                            reloadAction: {
+                                Task {
+                                    await viewModel.loadSounds()
+                                }
+                            }
+                        ),
+                        errorView: ErrorView(
+                            reactionNoLongerExists: viewModel.state == .reactionNoLongerExists,
+                            errorMessage: viewModel.errorMessage,
+                            tryAgainAction: {
+                                Task {
+                                    await viewModel.loadSounds()
+                                }
+                            }
                         )
-                        .frame(height: 260)
-                        .padding(.bottom, 6)
-                    },
-                    loadingView: LoadingView(),
-                    emptyStateView: EmptyStateView(
-                        reloadAction: {
-                            Task {
-                                await viewModel.loadSounds()
-                            }
-                        }
-                    ),
-                    errorView: ErrorView(
-                        reactionNoLongerExists: viewModel.state == .reactionNoLongerExists,
-                        errorMessage: viewModel.errorMessage,
-                        tryAgainAction: {
-                            Task {
-                                await viewModel.loadSounds()
-                            }
-                        }
                     )
-                )
-                .environment(TrendsHelper())
-            }
-            .toolbar {
-                ToolbarControls(
-                    soundSortOption: $viewModel.soundSortOption,
-                    playStopAction: { soundListViewModel.onPlayStopPlaylistSelected() },
-                    startSelectingAction: { soundListViewModel.onEnterMultiSelectModeSelected() },
-                    isPlayingPlaylist: soundListViewModel.isPlayingPlaylist,
-                    soundArrayIsEmpty: soundArrayIsEmpty,
-                    isSelecting: soundListViewModel.isSelectingSounds
-                )
-                .foregroundStyle(.white)
-                .opacity(toolbarControlsOpacity)
-                .disabled(soundArrayIsEmpty)
-                .onChange(of: viewModel.soundSortOption) {
-                    viewModel.sortSounds(by: viewModel.soundSortOption)
+                    .environment(TrendsHelper())
+                }
+                .toolbar {
+                    ToolbarControls(
+                        soundSortOption: $viewModel.soundSortOption,
+                        playStopAction: { soundListViewModel.onPlayStopPlaylistSelected() },
+                        startSelectingAction: { soundListViewModel.onEnterMultiSelectModeSelected() },
+                        isPlayingPlaylist: soundListViewModel.isPlayingPlaylist,
+                        soundArrayIsEmpty: soundArrayIsEmpty,
+                        isSelecting: soundListViewModel.isSelectingSounds
+                    )
+                    .foregroundStyle(.white)
+                    .opacity(toolbarControlsOpacity)
+                    .disabled(soundArrayIsEmpty)
+                    .onChange(of: viewModel.soundSortOption) {
+                        viewModel.sortSounds(by: viewModel.soundSortOption)
+                    }
+                }
+                .oneTimeTask {
+                    await viewModel.loadSounds()
+                }
+                .onAppear {
+                    Analytics().send(
+                        originatingScreen: "ReactionDetailView",
+                        action: "didViewReaction(\(viewModel.reaction.title))"
+                    )
                 }
             }
-            .oneTimeTask {
-                await viewModel.loadSounds()
-            }
-            .onAppear {
-                Analytics().send(
-                    originatingScreen: "ReactionDetailView",
-                    action: "didViewReaction(\(viewModel.reaction.title))"
-                )
-            }
+            .edgesIgnoringSafeArea(.top)
         }
-        .edgesIgnoringSafeArea(.top)
     }
 }
