+++ MedoDelirioBrasilia/Sources/Views/iPad/StandaloneFavoritesViewModel.swift
